move all x11 in dedicated folder
